Move lakeFS file system and file to top-level module
Adjust all example imports in the README, in tests, and the registration
hooks for `pyproject.toml`.
