+++ src/lakefs_spec/__init__.py
@@ -1,1 +1,3 @@
+from .spec import LakeFSFile, LakeFSFileSystem
+
 version = "0.1.1"
