+++ tests/test_put_file.py
@@ -1,6 +1,6 @@
 from lakefs_client.client import LakeFSClient
 
-from lakefs_spec.spec import LakeFSFileSystem
+from lakefs_spec import LakeFSFileSystem
 from tests.util import RandomFileFactory
